+++ android/src/main/kotlin/com/smileidentity/flutter/products/enhanceddocv/SmileIDEnhancedDocumentVerification.kt
@@ -8,6 +8,7 @@ import com.smileidentity.flutter.results.DocumentCaptureResult
 import com.smileidentity.flutter.utils.DocumentCaptureResultAdapter
 import com.smileidentity.flutter.views.SmileComposablePlatformView
 import com.smileidentity.flutter.views.SmileIDViewFactory
+import com.smileidentity.models.AutoCapture
 import com.smileidentity.models.ConsentInformation
 import com.smileidentity.models.ConsentedInformation
 import com.smileidentity.results.SmileIDResult
@@ -15,6 +16,8 @@ import com.smileidentity.util.randomJobId
 import com.smileidentity.util.randomUserId
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.platform.PlatformViewFactory
+import kotlin.time.Duration.Companion.milliseconds
+import kotlin.time.Duration.Companion.seconds
 import kotlinx.collections.immutable.toImmutableMap
 
 internal class SmileIDEnhancedDocumentVerification private constructor(
@@ -71,7 +74,11 @@ internal class SmileIDEnhancedDocumentVerification private constructor(
             captureBothSides = args["captureBothSides"] as? Boolean ?: true,
             userId = args["userId"] as? String ?: randomUserId(),
             jobId = args["jobId"] as? String ?: randomJobId(),
-            enableAutoCapture = args["enableAutoCapture"] as? Boolean ?: true,
+            autoCaptureTimeout = (args["autoCaptureTimeout"] as? Int)?.toLong()?.milliseconds
+                ?: 10.seconds,
+            autoCapture = (args["autoCapture"] as? String)?.lowercase()?.let { input ->
+                AutoCapture.entries.firstOrNull { it.name.lowercase() == input }
+            } ?: AutoCapture.AutoCapture,
             allowNewEnroll = args["allowNewEnroll"] as? Boolean ?: false,
             showAttribution = args["showAttribution"] as? Boolean ?: true,
             allowAgentMode = args["allowAgentMode"] as? Boolean ?: false,