+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * Removed `AntiFraud` response in `JobStatus` calls
+* Updated the `targetSdk` to 36 and updated the AGP version 
 
 ## 11.0.6 - July 11, 2025