+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Removed `AntiFraud` response in `JobStatus` calls
 * Updated the `targetSdk` to 36 and updated the AGP version 
+* Removed the default `ConsentInformation`
 
 ## 11.0.6 - July 11, 2025